ccl: move expression walking for IMPORT out of row converter and put it in exp_walker.go instead.
When supporting default expressions for selected non-immutable functions, we not
only have to walk through the default expression to make sure that each
function expression is explicitly supported, but also follow our own overrides.
These code are originally in `row_converter.go` and could be extremely long as
other default expressions are supported later. This PR tries to address this issue
by moving it to a new file expr_walker.go within the `row` package.

Release note: None
